Add Query Param Support for WebSockets
Add Query Param Support for Websockets:
* The support is added as a function of both HandShankeConnection struct and the Connection sruct. A package refactoring also has been done.
* Existing example has been modified to include the new change.

Resolve #4275
