+++ docs/ballerina-by-example/examples/websocket-chat-application/websocket-chat-application.bal
@@ -1,27 +1,34 @@
 import ballerina.net.ws;
 
 @ws:configuration {
-    basePath: "/chat/{fname}+{lname}/{age}",
+    basePath: "/chat/{name}",
     port:9090
 }
 service<ws> ChatApp {
 
     map consMap = {};
 
-    resource onOpen(ws:Connection conn, string fname, string lname, string age) {
+    resource onOpen(ws:Connection conn, string name) {
         consMap[conn.getID()] = conn;
-        string msg = string `{{fname}} {{lname}} with age {{age}} connected to chat`;
+        map params = conn.getQueryParams();
+        var age, err = (string)params.age;
+        string msg;
+        if (err == null) {
+            msg = string `{{name}} with age {{age}} connected to chat`;
+        } else {
+            msg = string `{{name}} connected to chat`;
+        }
         broadcast(consMap, msg);
     }
 
-    resource onTextMessage(ws:Connection con, ws:TextFrame frame, string fname) {
-        string msg = string `{{fname}}: {{frame.text}}`;
+    resource onTextMessage(ws:Connection con, ws:TextFrame frame, string name) {
+        string msg = string `{{name}}: {{frame.text}}`;
         println(msg);
         broadcast(consMap, msg);
     }
 
-    resource onClose(ws:Connection con, ws:CloseFrame frame, string fname) {
-        string msg = string `{{fname}} left the chat`;
+    resource onClose(ws:Connection con, ws:CloseFrame frame, string name) {
+        string msg = string `{{name}} left the chat`;
         consMap.remove(con.getID());
         broadcast(consMap, msg);
     }
