+++ stdlib/ballerina-http/src/main/java/org/ballerinalang/net/ws/nativeimpl/connection/GetParentConnection.java
@@ -16,44 +16,43 @@
  *  under the License.
  */
 
-package org.ballerinalang.net.ws.nativeimpl;
+package org.ballerinalang.net.ws.nativeimpl.connection;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.natives.AbstractNativeFunction;
-import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.Receiver;
+import org.ballerinalang.natives.annotations.ReturnType;
 import org.ballerinalang.net.ws.Constants;
-import org.wso2.transport.http.netty.contract.websocket.WebSocketInitMessage;
+import org.ballerinalang.net.ws.WebSocketConnectionManager;
+import org.ballerinalang.net.ws.WsOpenConnectionInfo;
 
 /**
- * Get the ID of the connection.
+ * Get parent connection is exists.
  *
  * @since 0.94
  */
 
 @BallerinaFunction(
         packageName = "ballerina.net.ws",
-        functionName = "cancelHandshake",
-        receiver = @Receiver(type = TypeKind.STRUCT, structType = "HandshakeConnection",
+        functionName = "getParentConnection",
+        receiver = @Receiver(type = TypeKind.STRUCT, structType = "Connection",
                              structPackage = "ballerina.net.ws"),
-        args = {@Argument(name = "statusCode", type = TypeKind.INT),
-                @Argument(name = "reason", type = TypeKind.STRING)},
+        returnType = {@ReturnType(type = TypeKind.STRUCT, structType = "Connection",
+                                  structPackage = "ballerina.net.ws")},
         isPublic = true
 )
-public class CancelHandshake extends AbstractNativeFunction {
+public class GetParentConnection extends AbstractNativeFunction {
 
     @Override
     public BValue[] execute(Context context) {
-        BStruct handshakeConnection = (BStruct) getRefArgument(context, 0);
-        int statusCode = (int) getIntArgument(context, 0);
-        String reason = getStringArgument(context, 0);
-        WebSocketInitMessage initMessage =
-                (WebSocketInitMessage) handshakeConnection.getNativeData(Constants.WEBSOCKET_MESSAGE);
-        initMessage.cancelHandShake(statusCode, reason);
-        return VOID_RETURN;
+        BStruct wsConnection = (BStruct) getRefArgument(context, 0);
+        String parentConnectionID = (String) wsConnection.getNativeData(Constants.NATIVE_DATA_PARENT_CONNECTION_ID);
+        WsOpenConnectionInfo connectionInfo =
+                WebSocketConnectionManager.getInstance().getConnectionInfo(parentConnectionID);
+        return getBValues(connectionInfo.getWsConnection());
     }
 }
