+++ stdlib/ballerina-http/src/main/java/org/ballerinalang/net/ws/nativeimpl/connection/GetQueryParams.java
@@ -16,54 +16,40 @@
  * under the License.
  */
 
-package org.ballerinalang.net.http.nativeimpl.inbound.request;
+package org.ballerinalang.net.ws.nativeimpl.connection;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BMap;
-import org.ballerinalang.model.values.BString;
-import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.natives.AbstractNativeFunction;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.Receiver;
 import org.ballerinalang.natives.annotations.ReturnType;
-import org.ballerinalang.net.http.Constants;
-import org.ballerinalang.net.uri.URIUtil;
+import org.ballerinalang.net.ws.WebSocketUtil;
 import org.ballerinalang.util.exceptions.BallerinaException;
-import org.wso2.transport.http.netty.message.HTTPCarbonMessage;
 
 /**
- * Get the Query params from HTTP message and return a map.
+ * Get the Query params from Websocket Connection and return a map.
  *
- * @since 0.94
+ * @since 0.961.0
  */
 @BallerinaFunction(
-        packageName = "ballerina.net.http",
+        packageName = "ballerina.net.ws",
         functionName = "getQueryParams",
-        receiver = @Receiver(type = TypeKind.STRUCT, structType = "InRequest",
-                             structPackage = "ballerina.net.http"),
+        receiver = @Receiver(type = TypeKind.STRUCT, structType = "Connection",
+                             structPackage = "ballerina.net.ws"),
         returnType = {@ReturnType(type = TypeKind.MAP, elementType = TypeKind.STRING)},
         isPublic = true
 )
 public class GetQueryParams extends AbstractNativeFunction {
+
     @Override
     public BValue[] execute(Context context) {
         try {
-            BStruct requestStruct  = ((BStruct) getRefArgument(context, 0));
-            HTTPCarbonMessage httpCarbonMessage = (HTTPCarbonMessage) requestStruct
-                    .getNativeData(Constants.TRANSPORT_MESSAGE);
-
-            if (httpCarbonMessage.getProperty(Constants.QUERY_STR) != null) {
-                String queryString = (String) httpCarbonMessage.getProperty(Constants.QUERY_STR);
-                BMap<String, BString> params = new BMap<>();
-                URIUtil.populateQueryParamMap(queryString, params);
-                return getBValues(params);
-            } else {
-                return getBValues(new BMap<>());
-            }
+            return WebSocketUtil.getQueryParams(context, this);
         } catch (Throwable e) {
-            throw new BallerinaException("Error while retrieving query param from message: " + e.getMessage());
+            throw new BallerinaException(
+                    "Error occurred while retrieving query parameters from Connection: " + e.getMessage());
         }
     }
 }
