+++ stdlib/ballerina-http/src/main/java/org/ballerinalang/net/ws/nativeimpl/connection/GetUpgradeHeader.java
@@ -16,42 +16,46 @@
  *  under the License.
  */
 
-package org.ballerinalang.net.ws.nativeimpl;
+package org.ballerinalang.net.ws.nativeimpl.connection;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BBoolean;
+import org.ballerinalang.model.values.BString;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.natives.AbstractNativeFunction;
+import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.Receiver;
 import org.ballerinalang.natives.annotations.ReturnType;
 import org.ballerinalang.net.ws.Constants;
 
-import javax.websocket.Session;
+import java.util.Locale;
+import java.util.Map;
 
 /**
- * Check whether the connection is secure connection or not.
+ * Get upgrade header for a given key.
  *
  * @since 0.94
  */
 
 @BallerinaFunction(
         packageName = "ballerina.net.ws",
-        functionName = "isSecure",
+        functionName = "getUpgradeHeader",
         receiver = @Receiver(type = TypeKind.STRUCT, structType = "Connection",
                              structPackage = "ballerina.net.ws"),
-        returnType = {@ReturnType(type = TypeKind.BOOLEAN)},
+        args = {@Argument(name = "text", type = TypeKind.STRING)},
+        returnType = {@ReturnType(type = TypeKind.STRING)},
         isPublic = true
 )
-public class IsSecure extends AbstractNativeFunction {
+public class GetUpgradeHeader extends AbstractNativeFunction {
 
     @Override
     public BValue[] execute(Context context) {
         BStruct wsConnection = (BStruct) getRefArgument(context, 0);
-        Session session = (Session) wsConnection.getNativeData(Constants.NATIVE_DATA_WEBSOCKET_SESSION);
-        boolean isSecuredConnection = session.isSecure();
-        return getBValues(new BBoolean(isSecuredConnection));
+        String key = getStringArgument(context, 0).toLowerCase(Locale.ENGLISH);
+        Map<String, String> upgradeHeaders =
+                (Map<String, String>) wsConnection.getNativeData(Constants.NATIVE_DATA_UPGRADE_HEADERS);
+        return getBValues(new BString(upgradeHeaders.get(key)));
     }
 }
