+++ stdlib/ballerina-http/src/main/java/org/ballerinalang/net/ws/nativeimpl/connection/GetUpgradeHeaders.java
@@ -16,10 +16,12 @@
  *  under the License.
  */
 
-package org.ballerinalang.net.ws.nativeimpl;
+package org.ballerinalang.net.ws.nativeimpl.connection;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
+import org.ballerinalang.model.values.BMap;
+import org.ballerinalang.model.values.BString;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.natives.AbstractNativeFunction;
@@ -27,32 +29,33 @@ import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.Receiver;
 import org.ballerinalang.natives.annotations.ReturnType;
 import org.ballerinalang.net.ws.Constants;
-import org.ballerinalang.net.ws.WebSocketConnectionManager;
-import org.ballerinalang.net.ws.WsOpenConnectionInfo;
+
+import java.util.Map;
 
 /**
- * Get parent connection is exists.
+ * Get all the upgrade headers.
  *
  * @since 0.94
  */
 
 @BallerinaFunction(
         packageName = "ballerina.net.ws",
-        functionName = "getParentConnection",
+        functionName = "getUpgradeHeaders",
         receiver = @Receiver(type = TypeKind.STRUCT, structType = "Connection",
                              structPackage = "ballerina.net.ws"),
-        returnType = {@ReturnType(type = TypeKind.STRUCT, structType = "Connection",
-                                  structPackage = "ballerina.net.ws")},
+        returnType = {@ReturnType(type = TypeKind.MAP)},
         isPublic = true
 )
-public class GetParentConnection extends AbstractNativeFunction {
+public class GetUpgradeHeaders extends AbstractNativeFunction {
 
     @Override
     public BValue[] execute(Context context) {
         BStruct wsConnection = (BStruct) getRefArgument(context, 0);
-        String parentConnectionID = (String) wsConnection.getNativeData(Constants.NATIVE_DATA_PARENT_CONNECTION_ID);
-        WsOpenConnectionInfo connectionInfo =
-                WebSocketConnectionManager.getInstance().getConnectionInfo(parentConnectionID);
-        return getBValues(connectionInfo.getWsConnection());
+        Map<String, String> upgradeHeaders =
+                (Map<String, String>) wsConnection.getNativeData(Constants.NATIVE_DATA_UPGRADE_HEADERS);
+        BMap<String, BString> bUpgradeHeaders = new BMap<>();
+        upgradeHeaders.entrySet().forEach(
+                upgradeHeader -> bUpgradeHeaders.put(upgradeHeader.getKey(), new BString(upgradeHeader.getValue())));
+        return getBValues(bUpgradeHeaders);
     }
 }
