+++ stdlib/ballerina-http/src/main/java/org/ballerinalang/net/ws/nativeimpl/connection/IsSecure.java
@@ -16,48 +16,42 @@
  *  under the License.
  */
 
-package org.ballerinalang.net.ws.nativeimpl;
+package org.ballerinalang.net.ws.nativeimpl.connection;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
+import org.ballerinalang.model.values.BBoolean;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.natives.AbstractNativeFunction;
-import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.Receiver;
+import org.ballerinalang.natives.annotations.ReturnType;
 import org.ballerinalang.net.ws.Constants;
-import org.ballerinalang.util.exceptions.BallerinaException;
 
-import java.nio.ByteBuffer;
 import javax.websocket.Session;
 
 /**
- * Push binary data to the other end of the connection.
+ * Check whether the connection is secure connection or not.
  *
  * @since 0.94
  */
 
 @BallerinaFunction(
         packageName = "ballerina.net.ws",
-        functionName = "pushBinary",
+        functionName = "isSecure",
         receiver = @Receiver(type = TypeKind.STRUCT, structType = "Connection",
                              structPackage = "ballerina.net.ws"),
-        args = {@Argument(name = "binaryData", type = TypeKind.BLOB)},
+        returnType = {@ReturnType(type = TypeKind.BOOLEAN)},
         isPublic = true
 )
-public class PushBinary extends AbstractNativeFunction {
+public class IsSecure extends AbstractNativeFunction {
 
     @Override
     public BValue[] execute(Context context) {
-        try {
-            BStruct wsConnection = (BStruct) getRefArgument(context, 0);
-            Session session = (Session) wsConnection.getNativeData(Constants.NATIVE_DATA_WEBSOCKET_SESSION);
-            byte[] binaryData = getBlobArgument(context, 0);
-            session.getBasicRemote().sendBinary(ByteBuffer.wrap(binaryData));
-        } catch (Throwable e) {
-            throw new BallerinaException("Cannot send the message. Error occurred.");
-        }
-        return VOID_RETURN;
+        BStruct wsConnection = (BStruct) getRefArgument(context, 0);
+        Session session = (Session) wsConnection.getNativeData(Constants.NATIVE_DATA_WEBSOCKET_SESSION);
+        boolean isSecuredConnection = session.isSecure();
+        return getBValues(new BBoolean(isSecuredConnection));
     }
 }
