+++ stdlib/ballerina-http/src/main/java/org/ballerinalang/net/ws/nativeimpl/connection/Pong.java
@@ -16,42 +16,48 @@
  *  under the License.
  */
 
-package org.ballerinalang.net.ws.nativeimpl;
+package org.ballerinalang.net.ws.nativeimpl.connection;
 
 import org.ballerinalang.bre.Context;
 import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.model.values.BString;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.natives.AbstractNativeFunction;
+import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinalang.natives.annotations.Receiver;
-import org.ballerinalang.natives.annotations.ReturnType;
 import org.ballerinalang.net.ws.Constants;
+import org.ballerinalang.util.exceptions.BallerinaException;
 
+import java.nio.ByteBuffer;
 import javax.websocket.Session;
 
 /**
- * Get negotiated sub protocol.
+ * Push binary data to the other end of the connection.
  *
  * @since 0.94
  */
 
 @BallerinaFunction(
         packageName = "ballerina.net.ws",
-        functionName = "getNegotiatedSubProtocol",
+        functionName = "pong",
         receiver = @Receiver(type = TypeKind.STRUCT, structType = "Connection",
                              structPackage = "ballerina.net.ws"),
-        returnType = {@ReturnType(type = TypeKind.STRING)},
+        args = {@Argument(name = "binaryData", type = TypeKind.BLOB)},
         isPublic = true
 )
-public class GetNegotiatedSubProtocol extends AbstractNativeFunction {
+public class Pong extends AbstractNativeFunction {
 
     @Override
     public BValue[] execute(Context context) {
-        BStruct wsConnection = (BStruct) getRefArgument(context, 0);
-        Session session = (Session) wsConnection.getNativeData(Constants.NATIVE_DATA_WEBSOCKET_SESSION);
-        String negotiatedSubProtocol = session.getNegotiatedSubprotocol();
-        return getBValues(new BString(negotiatedSubProtocol));
+        try {
+            BStruct wsConnection = (BStruct) getRefArgument(context, 0);
+            Session session = (Session) wsConnection.getNativeData(Constants.NATIVE_DATA_WEBSOCKET_SESSION);
+            byte[] binaryData = getBlobArgument(context, 0);
+            session.getBasicRemote().sendPong(ByteBuffer.wrap(binaryData));
+        } catch (Throwable e) {
+            throw new BallerinaException("Cannot send the message. Error occurred.");
+        }
+        return VOID_RETURN;
     }
 }
